requirements.txt
amqp==5.0.5
argon2-cffi==20.1.0
asgiref==3.3.1
async-generator==1.10
attrs==20.3.0
autobahn==20.12.3
Automat==20.2.0
Babel==2.8.0
backcall==0.2.0
billiard==3.6.3.0
bleach==3.2.2
celery==5.0.5
cffi==1.14.0
channels==3.0.3
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.4
constantly==15.1.0
cryptography==3.3.1
daphne==3.0.1
decorator==4.4.2
defusedxml==0.6.0
Django==3.1.5
django-celery-beat==2.2.0
django-cors-headers==3.4.0
django-countries==6.1.3
django-extensions==3.1.0
django-extra-fields==2.0.5
django-filter==2.3.0
django-js-asset==1.2.2
django-mptt==0.11.0
django-phonenumber-field==5.0.0
django-polymorphic==2.1.2
django-prices==2.2.0
django-rest-knox==4.1.0
django-reversion==3.0.7
django-simple-history==2.10.0
django-solo==1.1.5
django-templated-mail==1.1.1
django-timezone-field==4.1.1
djangorestframework==3.12.2
djangorestframework-recursive==0.1.2
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
drf-writable-nested==0.6.2
enmerkar==0.7.1
entrypoints==0.3
hyperlink==21.0.0
idna==3.1
importlib-metadata==1.6.1
incremental==17.5.0
ipykernel==5.4.3
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==2.11.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.11
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kombu==5.0.2
Markdown==3.2.2
MarkupSafe==1.1.1
mistune==0.8.4
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.4.3
notebook==6.2.0
numpy==1.19.4
packaging==20.8
pandocfilters==1.4.3
parso==0.8.1
phonenumbers==8.12.15
pickleshare==0.7.5
Pillow==7.1.2
prices==1.0.0
prometheus-client==0.9.0
prompt-toolkit==3.0.14
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.7.4
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-crontab==2.5.1
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.1
pywin32==300
pywinpty==0.5.7
pyzmq==21.0.1
qtconsole==5.0.2
QtPy==1.9.0
satchless==1.2.0
Send2Trash==1.5.0
service-identity==18.1.0
six==1.15.0
sqlparse==0.3.1
terminado==0.9.2
testpath==0.4.4
text-unidecode==1.3
tornado==6.1
traitlets==5.0.5
Twisted==20.3.0
txaio==20.12.1
typing==3.7.4.1
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==3.1.0
zope.interface==5.2.0
